Extract bundle insertion code
